requirements.txt
Adafruit-Blinka==8.6.0
adafruit-circuitpython-busdevice==5.2.3
adafruit-circuitpython-neopixel==6.3.6
adafruit-circuitpython-pixelbuf==1.1.8
adafruit-circuitpython-requests==1.12.9
adafruit-circuitpython-typing==1.8.2
Adafruit-PlatformDetect==3.32.0
Adafruit-PureIO==1.1.9
certifi==2023.7.22
charset-normalizer==3.2.0
Encoder==1.1
idna==3.4
playsound==1.3.0
pyftdi==0.54.0
pyserial==3.5
pyttsx3==2.90
pyusb==1.2.1
requests==2.31.0
rpi-ws281x==4.3.4
RPi.GPIO==0.7.1
rpimotorlib==3.2
smbus==1.1.post2
sysv-ipc==1.1.0
tabulate==0.9.0
tones==1.2.0
typing-extensions==4.4.0
urllib3==2.0.4
